refactor(layout): 페이지 레이아웃 재정비 (wrapper-container-section 순으로 작아짐)
- admin, apply-form 페이지 제외하고 적용
