+++ src/pages/item-list/ItemListPage.tsx
@@ -5,18 +5,16 @@ import { ItemListDescriptionSection, ItemListSection } from '@/features';
 
 import { BackButton, NAVIGATE_BUTTONS_DATA } from '@/shared';
 
-import { LogoContainer } from '@/widgets';
+import { ContentsContainer, PageWrapper } from '@/widgets';
 
 export const ItemListPage = () => {
   return (
-    <main className='flex h-full flex-col items-center justify-center gap-3 text-center'>
-      <LogoContainer title={NAVIGATE_BUTTONS_DATA[0].title}>
+    <PageWrapper>
+      <ContentsContainer title={NAVIGATE_BUTTONS_DATA[0].title}>
         <ItemListDescriptionSection />
         <ItemListSection />
-      </LogoContainer>
-      <div className='w-full pt-2 pb-6'>
-        <BackButton />
-      </div>
-    </main>
+      </ContentsContainer>
+      <BackButton />
+    </PageWrapper>
   );
 };
