+++ src/pages/my-item/MyItemPage.tsx
@@ -7,7 +7,7 @@ import {
 
 import { BackButton, NAVIGATE_BUTTONS_DATA, useGetMockData } from '@/shared';
 
-import { DescriptionSection, LogoContainer } from '@/widgets';
+import { ContentsContainer, DescriptionSection, PageWrapper } from '@/widgets';
 
 export const MyItemPage = () => {
   const { data: columnData, isPending } = useGetMockData(MY_ITEM_BODY);
@@ -16,8 +16,8 @@ export const MyItemPage = () => {
   const name = '김해달';
 
   return (
-    <div className='flex h-full flex-col items-center justify-center gap-3 text-center'>
-      <LogoContainer title={NAVIGATE_BUTTONS_DATA[2].title}>
+    <PageWrapper>
+      <ContentsContainer title={NAVIGATE_BUTTONS_DATA[2].title}>
         <DescriptionSection
           name={name}
           description={NAVIGATE_BUTTONS_DATA[2].description}
@@ -29,8 +29,8 @@ export const MyItemPage = () => {
           headerData={MY_ITEM_HEADERS}
           bodyData={MY_ITEM_BODY}
         />
-      </LogoContainer>
+      </ContentsContainer>
       <BackButton />
-    </div>
+    </PageWrapper>
   );
 };
