+++ src/widgets/container/index.ts
@@ -1,3 +1,3 @@
 export * from './AuthContainer';
 export * from './AuthLinkContainer';
-export * from './LogoContainer';
+export * from './ContentsContainer';
