+++ src/widgets/index.ts
@@ -1,2 +1,3 @@
 export * from './container';
 export * from './sections';
+export * from './wrapper';
